Add verifier GetBlockID (#23)
This allows to search for a block containing a specific fact, and
retrieve it's index. This is useful in proof of possession flow, where
the verifier needs to hash all the blocks before the one containing the
user signature.

Also fixed the `biscuit.SHA256Sum` function which were not including
the root key in the hash.