Refactor datalog engine (#98)
* clarify the datalog engine

* replace the Combinator struct with a gooutine and a channel

the goroutine generates combinations of facts, matches them with
predicates, run expressions and send valid ones over the channel. This
allows us to generate valid facts lazily and consume them as needed

* reactivate the samples test about symbols

this test was failing because the datalog engine was spending too much
time evaluating it. With the new engine it should work properly

* update golangci

* factor the index advancing code