samples: properly check generated facts and rules
This implementation supports only v3 biscuits, so it cannot handle
facts and rules that are scoped in non-authority blocks. Now that the
samples provide this information, the test suite can simply filter
out non-authority / authorizer facts and rules to compare the final
state of the world with the samples.

This allows to stop skipping some tests.