parser: correctly parse complex expression
The grammar handles precedence through precedence climbing. The layering
was broken, so it did not handle correctly expressions with mixed precedences.

It also put operators with different precedences at the same level (`&&` and `||`), and
allowed parsing chains of comparison operators, even though they are not associative.

Some issues in the lexer prevented it from parsing `&&` correctly.