parser: tokenize whole hex strings at once
The lexer used to emit a token for the `hex:` sigil, and then a token
for the hex string that came after. Unfortunately, the lexer did not
handle strings like `12ab` (digits, then letters) correctly. The solution
is to rely on the `hex:` sigil to parse the whole hex string directly.