datalog: relax constraints on `.contains()`
Contains can be called with a set argument: in that case it uses set
inclusion instead of element inclusion.

Contains can be called with heterogeneous arguments: if the argument to
contains is of a different type than the set elements, then contains
will return false, instead of erroring out. This is consistent with
how contains behaves in the rust and haskell implementations.