Refresh test suite (#120)
* Allow parsing whole blocks & authorizers

* Allow loading a parsed authorizer directly

* test(samples): update samples

* test(samples): automated conformance tests

* test(samples): compare parsed code with provided biscuits

This makes sure that the go parser is consistent with the rust parser

* test(samples): print sorted facts and rules in debug.World

This makes the output predictable and usable in tests

* test(samples): ensure consistency with samples authorization results

* Format datalog dates as UTC RFC3339 dates

Dates in datalog don't carry timezone information, so it's better to
always display them as UTC dates to avoid inconsistencies
(even though the instant they describe is unchanged)

* test(samples): only skip verifications where it is required

World consistency checks only need to be skipped in cases with
scoped rules and checks. The rest of the samples should be tested

* samples: update samples

* samples: properly check generated facts and rules

This implementation supports only v3 biscuits, so it cannot handle
facts and rules that are scoped in non-authority blocks. Now that the
samples provide this information, the test suite can simply filter
out non-authority / authorizer facts and rules to compare the final
state of the world with the samples.

This allows to stop skipping some tests.